FIX: older Qt doesn't have QVariant::toFloat()
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@2257 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/gui/fview/cmtkFusionViewApplication.cxx
@@ -229,7 +229,7 @@ void
 cmtk::FusionViewApplication
 ::changeZoom( QAction* action )
 {
-  this->m_ZoomFactor = action->data().toFloat();
+  this->m_ZoomFactor = static_cast<float>( action->data().toDouble() ); // older Qt doesn't have QVariant::toFloat()
   this->UpdateFixedImage();
   this->UpdateMovingImage();
 }
